requirements.txt
aenum==3.1.15
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
moneybox==0.0.35
pydantic==1.10.13
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
